Update kuwait.js
So I'm a beginner, haven't used GitHub before,
With help of AI, I did manage to get the code written to add my sponsors logos to the page, but I'm unable to upload the actual images in the repository as to add to the fork. Please let me know I need to do further. 
Really sorry for bothering you with this.